+++ src/main/java/ssu/eatssu/domain/user/presentation/docs/UserControllerDocs.java
@@ -11,18 +11,18 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Pageable;
 import ssu.eatssu.domain.auth.security.CustomUserDetails;
-import ssu.eatssu.domain.partnership.dto.PartnershipResponse;
+import ssu.eatssu.domain.partnership.dto.response.PartnershipResponse;
 import ssu.eatssu.domain.slice.dto.SliceResponse;
-import ssu.eatssu.domain.user.dto.DepartmentResponse;
-import ssu.eatssu.domain.user.dto.GetCollegeResponse;
-import ssu.eatssu.domain.user.dto.GetDepartmentResponse;
-import ssu.eatssu.domain.user.dto.LanguageResponse;
-import ssu.eatssu.domain.user.dto.LanguageUpdateRequest;
-import ssu.eatssu.domain.user.dto.MyMealReviewResponse;
-import ssu.eatssu.domain.user.dto.MyPageResponse;
-import ssu.eatssu.domain.user.dto.MyReviewDetail;
-import ssu.eatssu.domain.user.dto.NicknameUpdateRequest;
-import ssu.eatssu.domain.user.dto.UpdateDepartmentRequest;
+import ssu.eatssu.domain.user.dto.response.DepartmentResponse;
+import ssu.eatssu.domain.user.dto.response.GetCollegeResponse;
+import ssu.eatssu.domain.user.dto.response.GetDepartmentResponse;
+import ssu.eatssu.domain.user.dto.response.LanguageResponse;
+import ssu.eatssu.domain.user.dto.request.LanguageUpdateRequest;
+import ssu.eatssu.domain.user.dto.response.MyMealReviewResponse;
+import ssu.eatssu.domain.user.dto.response.MyPageResponse;
+import ssu.eatssu.domain.user.dto.response.MyReviewDetail;
+import ssu.eatssu.domain.user.dto.request.NicknameUpdateRequest;
+import ssu.eatssu.domain.user.dto.request.UpdateDepartmentRequest;
 import ssu.eatssu.global.handler.response.BaseResponse;
 
 import java.util.List;
