+++ src/main/java/ssu/eatssu/domain/user/service/UserService.java
@@ -14,14 +14,14 @@ import ssu.eatssu.domain.user.department.entity.College;
 import ssu.eatssu.domain.user.department.entity.Department;
 import ssu.eatssu.domain.user.department.persistence.CollegeRepository;
 import ssu.eatssu.domain.user.department.persistence.DepartmentRepository;
-import ssu.eatssu.domain.user.dto.DepartmentResponse;
-import ssu.eatssu.domain.user.dto.GetCollegeResponse;
-import ssu.eatssu.domain.user.dto.GetDepartmentResponse;
-import ssu.eatssu.domain.user.dto.LanguageResponse;
-import ssu.eatssu.domain.user.dto.LanguageUpdateRequest;
-import ssu.eatssu.domain.user.dto.MyPageResponse;
-import ssu.eatssu.domain.user.dto.NicknameUpdateRequest;
-import ssu.eatssu.domain.user.dto.UpdateDepartmentRequest;
+import ssu.eatssu.domain.user.dto.response.DepartmentResponse;
+import ssu.eatssu.domain.user.dto.response.GetCollegeResponse;
+import ssu.eatssu.domain.user.dto.response.GetDepartmentResponse;
+import ssu.eatssu.domain.user.dto.response.LanguageResponse;
+import ssu.eatssu.domain.user.dto.request.LanguageUpdateRequest;
+import ssu.eatssu.domain.user.dto.response.MyPageResponse;
+import ssu.eatssu.domain.user.dto.request.NicknameUpdateRequest;
+import ssu.eatssu.domain.user.dto.request.UpdateDepartmentRequest;
 import ssu.eatssu.domain.user.entity.DeviceType;
 import ssu.eatssu.domain.user.entity.Language;
 import ssu.eatssu.domain.user.entity.User;
